roachtest: add CreateNewUser helper
This change adds a convenience helper that creates
a new user with a password if specified. It also
creates client certificates for the user.

This helper will simplify the process of creating a user
in roachtests, as well as allow us to keep track of
which users have been created. This will set up non default
user usage of AuthUserCert authentication later on.