roachtest: string match for transient errors as a fallback
The `require` package is commonly used through roachtest to
assert that no error occured. i.e. `require.NoError(t, err)`
However, this function does not preserve the error object. This
causes our transient error flake detection to not work. Since
`require` is an upstream dependency, we cannot easily change this.

This change adds a fallback to our flake detection that string
matches for the `TRANSIENT_ERROR` message we add. If found it
will mark the error as a flake to reduce noise.

However, we have seen other cases where we do not preserve the
error object but the code lives somewhere that is easily changeable for us.
In those cases, we ideally should fix the code instead of resorting
to this fallback.

To make sure we still do that, the fallback also explicity checks for
a message that `require.NoError` prepends to all errors. If we find
additional cases that require this fallback, we can review and add
them on a case by case basis.