roachprod: remove ServiceMode from dns registry
As of #125393, we no longer register shared-process tenants
in DNS records. This was a optimization made as they share
the same ports as the system tenant so no registration is
needed.

This change removes the concept of ServiceMode from the DNS
registry as we know that all registered services are either
external separate process services, or system tenants which
do not fall under either external or shared.